refactor combined_queries into separate files (no functionality changes)
no more solana.fm so no need to have a dedicated bitquery.ts/combined_queries.ts
file

split by liquidity pool/tracked token functionality instead
